[dev] feat: 조건부 라우팅 구현
디데이가 0일때 페이지 라우트!
