+++ src/components/BirthdayPage.tsx
@@ -1,0 +1,9 @@
+function BirthdayPage() {
+    return (
+        <div className="text-center">
+            <h1 className="text-3xl font-bold">🎉 생일축하해요! 🎉</h1>
+        </div>
+    );
+}
+
+export default BirthdayPage;
